pyproject.toml
[project]
name = "osrs-voiceover-generator"
version = "0.1.0"
description = "generates voice over files by scrapping the osrs wiki for quest transcripts, and using the Elevenlabs SDK to generate voiceovers for them."
authors = [{name = "Kevin Orel Edry", email = "techg9@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "ffmpeg>=1.4",
    "elevenlabs>=1.3.1",
    "pydub>=0.25.1",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    "ffmpeg-python>=0.2.0",
    "prompt-toolkit>=3.0.47",
    "tqdm>=4.66.4",
    "python-dotenv>=1.0.1",
    # KokoroTTS dependencies (free local TTS)
    "kokoro>=0.9.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["voiceover_cli"]
